pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.5"
name = "phomemo-p12-tools"
description='Phomemo P12 printing tools'
readme = "README.md"
license = {file = "LICENSE"}
authors = [
	{name = 'Mirco Janisch', email='mircojanisch@hotmail.com'},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: System :: Hardware :: Hardware Drivers"
]
dependencies = [
    "Pillow",
    "pycairo",
    "pyserial",
]

[project.urls]
homepage='https://github.com/R3tr0BoiDX/phomemo-p12'
repository='https://github.com/R3tr0BoiDX/phomemo-p12'

[tool.setuptools.packages.find]
exclude = ["build", "tests"]
